Create README.md
Added description for Fraud Neuron framework, including details on how it complements MITRE ATT&CK, the data structure, and how to access and contribute. Provided information on the platform URL (neuron.axur.com) for accessing the framework and guidelines for submitting Pull Requests with changes to the `dataset.json` file.